Module quote_bot::db::actions

Structs

ListParams

Functions

create_quote
Creates a new quote and saves it into the database.

delete_quote

find_author
Finds an author in the database, if they do not exist, create a new one.

find_contains_quotes
Searches for a max of 5 quotes that contain the given string query.

find_listed_quotes

find_quote
Find a quote, return Some(Quote) if found or None, else an error.

find_rand_quote
Finds a random quote from the user.