Struct quote_bot::db::models::Quote

pub struct Quote {
    pub id: i32,
    pub message_id: String,
    pub quote: String,
    pub created_by_id: String,
    pub quoted_by_id: String,
    pub guild_id: String,
}

Fields
id: i32

Autogen SERIAL PRIMARY KEY

message_id: String

The discord message snowflake id, UNIQUE

quote: String

Quote body, with formatting.

created_by_id: String

Who created the original message, the person who was quoted.

quoted_by_id: String

The person who starred the message, the person who quoted.

guild_id: String